// api/src/server/twilio/types/interfaces.d.ts
// This file was auto-generated by prisma-generator-typescript-interfaces

export type PropertyOccupancy = "VACANT" | "OWNER" | "TENANT";

export interface RequestService {
  RequestServiceID: number;
  Description: string;
  Order?: Order[];
}

export interface Customer {
  CustomerID: string;
  Title: string | null;
  Name: string;
  FirstName: string;
  MiddleName: string | null;
  LastName: string;
  Suffix: string | null;
  PhoneNumber: string;
  EmailAddress: string | null;
  Brokerage: string | null;
  Orders?: Order[];
  newOrder?: Order;
}

export interface Order {
  OrderID: string;
  CreatedAt: Date;
  PropertyAddress: string;
  PropertyCounty: string | null;
  RequestedService?: RequestService;
  RequestedServiceID: number;
  RequestedInstallDate: Date;
  RequestedRemoveDate: Date | null;
  Occupancy: PropertyOccupancy;
  belongsTo?: Customer;
  CustomerID: string;
  Fullfillment?: Fullfillment | null;
  Attachments?: Attachment[];
  Remove?: boolean;
  isConfirmed: boolean;
}

export interface Employee {
  EmployeeID: string;
  FirstName: string;
  LastName: string;
  Fullfillment?: Fullfillment[];
}

export interface Fullfillment {
  Order?: Order;
  FullfilledOrderID: string;
  FullfilledAt: Date;
  EmployeeWhoFullfilled?: Employee;
  FullfilledEmployeeID: string;
}

export interface Attachment {
  Order?: Order;
  OrderID: string;
  AttachmentData: Uint8Array;
}